# py/malloc: Fix failed memory allocations by more aggressively freeing buffers.