--- a/ports/rp2/mpthreadport.h
+++ b/ports/rp2/mpthreadport.h
@@ -26,9 +26,10 @@
 #ifndef MICROPY_INCLUDED_RP2_MPTHREADPORT_H
 #define MICROPY_INCLUDED_RP2_MPTHREADPORT_H
 
-#include "pico/mutex.h"
+#include "mutex_extra.h"
 
 typedef struct mutex mp_thread_mutex_t;
+typedef recursive_mutex_nowait_t mp_thread_recursive_mutex_t;
 
 extern void *core_state[2];
 
@@ -63,6 +64,23 @@ static inline int mp_thread_mutex_lock(mp_thread_mutex_t *m, int wait) {
 
 static inline void mp_thread_mutex_unlock(mp_thread_mutex_t *m) {
     mutex_exit(m);
+}
+
+static inline void mp_thread_recursive_mutex_init(mp_thread_recursive_mutex_t *m) {
+    recursive_mutex_nowait_init(m);
+}
+
+static inline int mp_thread_recursive_mutex_lock(mp_thread_recursive_mutex_t *m, int wait) {
+    if (wait) {
+        recursive_mutex_nowait_enter_blocking(m);
+        return 1;
+    } else {
+        return recursive_mutex_nowait_try_enter(m, NULL);
+    }
+}
+
+static inline void mp_thread_recursive_mutex_unlock(mp_thread_recursive_mutex_t *m) {
+    recursive_mutex_nowait_exit(m);
 }
 
 #endif // MICROPY_INCLUDED_RP2_MPTHREADPORT_H