@dpgeorge 2024-12-03
This looks really clean @projectgus !  Thanks for the thorough implementation and testing.

That said, I can't help but think of the alternative where `gc_free()` could possibly be made to work when the GC is locked.  That would solve the bare-metal case without needing a list of pending tracked nodes to free, and also maybe improve other cases where `gc_free()` doesn't work if the heap is locked.

It could be tricky to implement though, eg it probably needs a recursive mutex for `GC_ENTER()`.  Or, could have a flag that's set when running a finaliser (eg `gc_finaliser_active = true`) and if this flag is set when `gc_free()` is called it runs only this bit of the function:
```c
    // free head and all of its tail blocks
    do {
        ATB_ANY_TO_FREE(area, block);
        block += 1;
    } while (ATB_GET_KIND(area, block) == AT_TAIL);
```
That would be more of a self-contained fix, contained to `gc.c`.