@projectgus 2024-12-03
>  `gc_free()` could possibly be made to work when the GC is locked

I think I started out with the same idea, and came to a sticking point that I couldn't seem to resolve. Unfortunately I no longer remember what that sticking point was, but I'll take a quick look and get back to you.