@projectgus 2024-12-04
OK, maybe I didn't try this approach befote... I just implemented it and it works better than I expected!

Currently fails on ports with `MICROPY_PY_THREAD && !MICROPY_PY_THREAD_GIL` as `GC_ENTER` isn't recursive yet, that needs updating and is a bit of work as different on each port.

Otherwise, seems like `gc_free()` can run unmodified in this case. It updates `gc_last_free_area` and `gc_last_free_atb_index` but these are both reset immediately after `gc_sweep` so changing them here seems pretty harmless.

### Running finalisers early

I had the same problem of potential for double-frees during `gc_sweep_all()`, so decided to try splitting out finalisers into an earlier pass as suggested above. Then I went back and optimised the pass a little bit for speed.

This seems to have turned out great!

Code size impact is small (this is the biggest code size impact in this PR, but the total size impact is *less* than the previous approach and we get the "run finalisers early on soft reset" behaviour as a bonus!)

| Iteration | qemu ARM text size | uni minimal text size |
| ------------ | -------------------------- | --------------- |
| baseline | 182052 | 175563 |
| split finaliser pass | 182072 | 175675 |
| optimised split finaliser pass | 182072 (!) | 175731 |

Garbage collecting large number of objects is also faster with this change, due to the rapid skipping of finaliser checks.  Ran [this test program](https://gist.github.com/projectgus/56af097dbe55898fda0bb18064b749c0) on ESP32_GENERIC+SPIRAM and ESP32_GENERIC_S3 (assuming ESP32+SPIRAM is close to worst case due to lots of slow RAM).

| Variant + Test | GC Time (master) | GC Time (this PR) | Delta |
| -------------------- | ----------- | ---- | ---- |
| ESP32_GENERIC+SPIRAM | | | |
| Baseline | 666us | 685us | +19us |
| objects | 477149us | 476484us | -665us |
| objects collected | 42260us | -8514us |
| ESP32_GENERIC_S3 | | | |
| Baseline | 281us | 274us | -7us |
| objects | 227321us | 226257us | -1064us |
| objects collected | 25271us | 18445us | -6826us |

I was going to make another test where there were thousands of objects with finalisers, but then I realised this doesn't really happen... all of the objects with finalisers are relatively limited system resources so at most there could be maybe a couple of dozen objects with finalisers live in the system. So optimising for a case where most blocks don't contain finalisers seems to be fine!