@projectgus 2024-12-10
This should be ready for review, have updated the PR description with the new approach and the results of `run-perfbench.py` (quite good!)